requirements/testing.txt
-r default.txt

pytest
pytest-aiohttp
pytest-asyncio
